feat(home-v2): cutover — / now renders HomeV2, legacy moved to /home-legacy
- Renamed src/pages/home → src/pages/homeLegacy via git mv
- Updated all 8 importers of pages/home/data to pages/homeLegacy/data
  (routes/baseRoutes, routes/heroRoutes, pages/dash/data,
   organisms/header, organisms/map, homeV2/adapters ×3,
   homeV2/sections/MapSection/MapFilterCard)
- PlatformRoutes: replaced eager Home import with lazy Home (homeV2)
  + lazy HomeLegacy (homeLegacy); / → HomeV2, /home-legacy → HomeLegacy;
  removed /home-v2 route
- Added HOME_LEGACY_PATH = "/home-legacy" to routes/path.ts
