+++ src/pages/homeV2/sections/NewsSection/NewsFeaturedCard.tsx
@@ -20,7 +20,7 @@ export function NewsFeaturedCard({ item }: { item: NewsItem }) {
         CAT_BG[item.category],
       ].join(" ")}
     >
-      <div className="absolute inset-0 p-6 md:p-8 flex flex-col justify-end text-white">
+      <div className="absolute bg-green inset-0 p-6 md:p-8 flex flex-col justify-end text-white">
         <span className="self-start mb-3 text-[10px] tracking-widest font-bold uppercase bg-white/20 backdrop-blur px-3 py-1 rounded-full">
           {item.category}
         </span>
