+++ src/routes/PlatformRoutes.tsx
@@ -37,7 +37,6 @@ import DashQuestionNew from "../pages/dashQuestionNew";
 import DashRoles from "../pages/dashRoles";
 import DashSimulado from "../pages/dashSimulado";
 import Forgot from "../pages/forgot";
-import Home from "../pages/home";
 import Login from "../pages/login";
 import Logout from "../pages/logout";
 import MainSimulate from "../pages/mainSimulate";
@@ -54,7 +53,7 @@ import { useAuthStore } from "../store/auth";
 import { BaseRoutes } from "./baseRoutes";
 import { HeroRoutes } from "./heroRoutes";
 import {
-  HOME_V2_PATH,
+  HOME_LEGACY_PATH,
   ACCOUNT_PATH,
   CONFIRM_EMAIL,
   CONTENT,
@@ -102,7 +101,8 @@ import {
 import ProtectedRoute from "./protectedRoute";
 import ProtectedRoutePermission from "./protectedRoutePermission";
 
-const HomeV2 = lazy(() => import("../pages/homeV2"));
+const Home = lazy(() => import("../pages/homeV2"));
+const HomeLegacy = lazy(() => import("../pages/homeLegacy"));
 
 export function PlatformRoutes() {
   const { data } = useAuthStore();
@@ -114,16 +114,23 @@ export function PlatformRoutes() {
       <Route element={<ConfirmEmailPage />} path={CONFIRM_EMAIL} />
       <Route element={<BaseRoutes />}>
         <Route element={<HeroRoutes />}>
-          <Route path={HOME_PATH} element={<Home />} />
-          <Route path={NEWS} element={<NewsPage />} />
           <Route
-            path={HOME_V2_PATH}
+            path={HOME_PATH}
             element={
               <Suspense fallback={null}>
-                <HomeV2 />
+                <Home />
               </Suspense>
             }
           />
+          <Route
+            path={HOME_LEGACY_PATH}
+            element={
+              <Suspense fallback={null}>
+                <HomeLegacy />
+              </Suspense>
+            }
+          />
+          <Route path={NEWS} element={<NewsPage />} />
         </Route>
         <Route path={LOGIN_PATH} element={<Login />} />
         <Route path={FORGOT_PASSWORD_PATH} element={<Forgot />} />
