+++ src/routes/path.ts
@@ -47,3 +47,4 @@ export const ESSAY_REVIEW_LIST = "revisao-redacoes";
 export const ESSAY_REVIEW_CURSINHO = "revisao-redacoes-cursinho";
 
 export const HOME_V2_PATH = "/home-v2";
+export const HOME_LEGACY_PATH = "/home-legacy";
